Error Fixing

+++ src/main/java/com/base/basesetup/entity/ChargerCostDebitNoteVO.java
@@ -59,7 +59,7 @@ public class ChargerCostDebitNoteVO {
 	@Column(name = "qty")
 	private int qty;
 	@Column(name = "gstpercent")
-	private Float GSTPercent;
+	private double GSTPercent;
 //	@Column(name = "tlcamount", precision = 10, scale = 2)
 //	private BigDecimal tlcAmount;
 	@Column(name = "gstamount", precision = 10, scale = 2)
